move s3 utilities into new s3_utils module, update submit_jobs in ndmg_cloud to use boto3
